@@ -55,115 +55,121 @@
     "node": ">=6.9.0",
     "npm": ">=3.10.3"
   },
-  "devDependencies": {
-    "@angular/compiler-cli": "^5.2.3",
-    "@angular/language-service": "^5.2.3",
-    "@angular/platform-server": "^5.2.3",
+  "optionalDependencies": {
     "@bahmutov/add-typescript-to-cypress": "^2.0.0",
-    "@compodoc/compodoc": "^1.0.5",
-    "@types/async": "^2.0.47",
-    "@types/browser-sync": "^0.0.38",
-    "@types/express": "^4.11.1",
-    "@types/gulp": "^4.0.5",
-    "@types/gulp-filter": "^3.0.31",
-    "@types/gulp-htmlmin": "^1.3.31",
-    "@types/gulp-load-plugins": "^0.0.31",
-    "@types/gulp-sass": "^0.0.31",
-    "@types/gulp-util": "^3.0.34",
     "@types/jasmine": "^2.8.6",
-    "@types/node": "^9.3.0",
-    "@types/rimraf": "2.0.2",
-    "@types/run-sequence": "^0.0.30",
-    "@types/systemjs": "^0.20.6",
-    "@types/yargs": "^11.0.0",
-    "async": "^2.6.0",
-    "autoprefixer": "^8.0.0",
-    "browser-sync": "^2.23.6",
-    "codelyzer": "^4.1.0",
-    "connect-history-api-fallback": "^1.5.0",
-    "cssnano": "^3.10.0",
-    "cypress": "^2.0.0",
-    "deep-extend": "^0.5.0",
-    "event-stream": "^3.3.4",
-    "express": "~4.16.2",
-    "express-history-api-fallback": "^2.2.1",
-    "gulp": "^3.9.1",
-    "gulp-cached": "^1.1.1",
-    "gulp-cheerio": "^0.6.3",
-    "gulp-concat": "^2.6.1",
-    "gulp-concat-css": "^3.0.0",
-    "gulp-filter": "^5.1.0",
-    "gulp-htmlmin": "^4.0.0",
-    "gulp-inject": "^4.3.0",
-    "gulp-inline-ng2-template": "^4.1.0",
-    "gulp-load-plugins": "^1.5.0",
-    "gulp-plumber": "~1.2.0",
-    "gulp-postcss": "^7.0.1",
-    "gulp-progeny": "^0.4.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-replace": "^0.6.1",
-    "gulp-sass": "^3.1.0",
-    "gulp-sourcemaps": "2.6.4",
-    "gulp-template": "^5.0.0",
-    "gulp-tslint": "^8.1.2",
-    "gulp-typescript": "^4.0.0",
-    "gulp-uglify": "^3.0.0",
-    "gulp-util": "^3.0.8",
-    "gulp-watch": "^5.0.0",
-    "is-ci": "^1.1.0",
-    "isstream": "^0.1.2",
-    "jasmine-core": "~2.99.0",
+    "cypress": "^3.0.1",
+    "jasmine-core": "^3.1.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage": "^1.1.1",
     "karma-jasmine": "^1.1.1",
     "karma-mocha-reporter": "^2.2.5",
-    "karma-remap-istanbul": "^0.6.0",
+    "karma-remap-istanbul": "^0.6.0"
+  },
+  "devDependencies": {
+    "@angular/compiler-cli": "^6.0.1",
+    "@angular/language-service": "^6.0.1",
+    "@angular/platform-server": "^6.0.1",
+    "@compodoc/compodoc": "^1.1.1",
+    "@types/async": "^2.0.48",
+    "@types/browser-sync": "^0.0.41",
+    "@types/express": "^4.11.1",
+    "@types/gulp": "^4.0.5",
+    "@types/gulp-filter": "^3.0.32",
+    "@types/gulp-htmlmin": "^1.3.32",
+    "@types/gulp-load-plugins": "^0.0.31",
+    "@types/gulp-sass": "^0.0.31",
+    "@types/gulp-util": "^3.0.34",
+    "@types/node": "^10.3.5",
+    "@types/rimraf": "2.0.2",
+    "@types/run-sequence": "^0.0.30",
+    "@types/systemjs": "^0.20.6",
+    "@types/yargs": "^11.0.0",
+    "async": "^2.6.0",
+    "autoprefixer": "^8.2.0",
+    "browser-sync": "^2.23.6",
+    "codelyzer": "^4.2.1",
+    "connect-history-api-fallback": "^1.5.0",
+    "cssnano": "^4.0.0",
+    "deep-extend": "^0.6.0",
+    "event-stream": "^3.3.4",
+    "express": "~4.16.3",
+    "express-history-api-fallback": "^2.2.1",
+    "gulp": "^3.9.1",
+    "gulp-cached": "^1.1.1",
+    "gulp-cheerio": "^0.6.3",
+    "gulp-concat": "^2.6.1",
+    "gulp-concat-css": "^3.1.0",
+    "gulp-filter": "^5.1.0",
+    "gulp-htmlmin": "^4.0.0",
+    "gulp-inject": "^4.3.2",
+    "gulp-inline-ng2-template": "^5.0.0",
+    "gulp-load-plugins": "^1.5.0",
+    "gulp-plumber": "~1.2.0",
+    "gulp-postcss": "^7.0.1",
+    "gulp-progeny": "^0.4.1",
+    "gulp-rename": "^1.2.2",
+    "gulp-replace": "^1.0.0",
+    "gulp-sass": "^4.0.0",
+    "gulp-sourcemaps": "2.6.4",
+    "gulp-template": "^5.0.0",
+    "gulp-tslint": "^8.1.2",
+    "gulp-typescript": "^4.0.2",
+    "gulp-uglify": "^3.0.0",
+    "gulp-util": "^3.0.8",
+    "gulp-watch": "^5.0.0",
+    "is-ci": "^1.1.0",
+    "isstream": "^0.1.2",
     "merge-stream": "^1.0.1",
     "minimatch": "^3.0.4",
+    "node-sass-magic-importer": "^5.1.3",
     "open": "0.0.5",
-    "remap-istanbul": "^0.10.0",
+    "remap-istanbul": "^0.11.0",
     "rimraf": "^2.6.2",
-    "rollup": "~0.56.0",
+    "rollup": "~0.62.0",
     "rollup-plugin-commonjs": "^9.0.0",
-    "rollup-plugin-includepaths": "0.2.2",
-    "rollup-plugin-node-resolve": "^3.0.2",
+    "rollup-plugin-includepaths": "0.2.3",
+    "rollup-plugin-node-resolve": "^3.3.0",
     "run-sequence": "^2.2.1",
     "semver": "^5.5.0",
     "serve-static": "^1.13.1",
-    "slash": "~1.0.0",
+    "slash": "~2.0.0",
     "source-map-explorer": "^1.5.0",
-    "supports-color": "^5.1.0",
-    "systemjs-builder": "0.16.12",
+    "supports-color": "^5.3.0",
+    "systemjs-builder": "0.16.13",
     "temp": "^0.8.3",
     "tildify": "^1.2.0",
     "traceur": "^0.0.111",
-    "ts-node": "^5.0.0",
+    "ts-node": "^6.0.0",
     "tslint": "^5.9.1",
-    "typescript": "^2.7.1",
-    "walk": "^2.3.9",
-    "yargs": "^11.0.0"
+    "typescript": "2.7.x",
+    "walk": "^2.3.13",
+    "yargs": "^12.0.1"
   },
   "dependencies": {
-    "@angular-devkit/build-optimizer": "^0.4.0",
-    "@angular/animations": "^5.2.3",
-    "@angular/common": "^5.2.3",
-    "@angular/compiler": "^5.2.3",
-    "@angular/core": "^5.2.3",
-    "@angular/forms": "^5.2.3",
-    "@angular/platform-browser": "^5.2.3",
-    "@angular/platform-browser-dynamic": "^5.2.3",
-    "@angular/router": "^5.2.3",
-    "@angular/service-worker": "^5.2.3",
-    "@angular/material": "^5.0.1",
-    "@angular/cdk": "^5.0.1",
+    "@angular-devkit/build-optimizer": "^0.6.1",
+    "@angular/animations": "^6.0.1",
+    "@angular/common": "^6.0.1",
+    "@angular/compiler": "^6.0.1",
+    "@angular/core": "^6.0.1",
+    "@angular/forms": "^6.0.1",
+    "@angular/platform-browser": "^6.0.1",
+    "@angular/platform-browser-dynamic": "^6.0.1",
+    "@angular/router": "^6.0.1",
+    "@angular/service-worker": "^6.0.1",
+    "@angular/material": "^6.3.3",
+    "@angular/cdk": "^6.3.3",
     "@types/hammerjs": "^2.0.31",
-    "core-js": "^2.5.3",
+    "core-js": "^2.5.4",
     "intl": "^1.2.5",
-    "rxjs": "^5.5.3",
-    "systemjs": "0.21.0",
-    "zone.js": "0.8.20",
+    "rxjs": "^6.1.0",
+    "systemjs": "0.21.4",
+    "zone.js": "~0.8.26"
+  },
+  "resolutions": {
+    "@types/sinon": "<5.0.0",
     "reflect-metadata": "^0.1.8",
     "hammerjs": "^2.0.8"
   }
