@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-// import 'rxjs/add/operator/do';  // for debugging
+
+import { Observable, of } from 'rxjs';
+import { catchError, tap } from 'rxjs/operators';
 
 /**
  * This class provides the NameList service with methods to read names and add names.
  */
-@Injectable()
+@Injectable({ providedIn: 'root' })
 export class NameListService {
 
   /**
@@ -21,9 +22,10 @@ export class NameListService {
    * @return {string[]} The Observable for the HTTP request.
    */
   get(): Observable<string[]> {
-    return this.http.get('assets/data.json')
-    //              .do(data => console.log('server data:', data))  // debug
-                    .catch(this.handleError);
+    return this.http.get<string[]>('assets/data.json')
+                    .pipe(
+    //                tap((data: string[]) => console.log('server data:', data)), // debug
+                      catchError(this.handleError));
   }
 
   /**
@@ -35,7 +37,8 @@ export class NameListService {
     const errMsg = (error.message) ? error.message :
       error.status ? `${error.status} - ${error.statusText}` : 'Server error';
     console.error(errMsg); // log to console instead
-    return Observable.throw(errMsg);
+
+    return of(errMsg);
   }
 }
 
