@@ -1,10 +1,12 @@
 describe('About', () => {
 
-  beforeEach(() => {
+  it('should have correct feature heading', () => {
     cy.visit('/about');
+    cy.get('mat-card h2').should('have.text', 'Features');
   });
 
-  it('should have correct feature heading', () => {
+  it('should use our custom cypress command', () => {
+    cy.visitAbout();
     cy.get('mat-card h2').should('have.text', 'Features');
   });
 
