@@ -2,12 +2,10 @@ import { NgModule } from '@angular/core';
 import { HomeComponent } from './home.component';
 import { HomeRoutingModule } from './home-routing.module';
 import { SharedModule } from '../shared/shared.module';
-import { NameListService } from '../shared/name-list/name-list.service';
 
 @NgModule({
   imports: [HomeRoutingModule, SharedModule],
   declarations: [HomeComponent],
-  exports: [HomeComponent],
-  providers: [NameListService]
+  exports: [HomeComponent]
 })
 export class HomeModule { }
