@@ -8,7 +8,24 @@ System.config({
     '': {
       defaultExtension: 'js'
     },
-    rxjs: {
+    'rxjs': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    'rxjs/ajax': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    'rxjs/operators': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    'rxjs/testing': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    'rxjs/webSocket': {
+      main: 'index.js',
       defaultExtension: 'js'
     }
   }
