@@ -2,6 +2,8 @@ import { join } from 'path';
 import * as slash from 'slash';
 import { argv } from 'yargs';
 
+const magicImporter = require('node-sass-magic-importer');
+
 import {
   BuildType,
   ExtendPackages,
@@ -427,12 +429,6 @@ export class SeedConfig {
       inject: 'shims',
       buildType: BUILD_TYPES.DEVELOPMENT
     },
-    // Temporary fix. See https://github.com/angular/angular/issues/9359
-    {
-      src: '.tmp/Rx.min.js',
-      inject: 'libs',
-      buildType: BUILD_TYPES.DEVELOPMENT
-    }
   ];
 
   /**
@@ -477,6 +473,12 @@ export class SeedConfig {
    * @type {any}
    */
   SYSTEM_CONFIG_DEV: any = {
+    bundles: {
+      'node_modules/.tmp/Rx.min.js': [
+        'rxjs',
+        'rxjs/*'
+      ]
+    },
     paths: {
       [this.BOOTSTRAP_MODULE]: `${this.APP_BASE}${this.BOOTSTRAP_MODULE}`,
       '@angular/animations':
@@ -512,7 +514,6 @@ export class SeedConfig {
         'node_modules/@angular/platform-browser-dynamic/bundles/platform-browser-dynamic-testing.umd.js',
       '@angular/router/testing':
         'node_modules/@angular/router/bundles/router-testing.umd.js',
-      'rxjs/operators': 'node_modules/rxjs/operators',
 
       'app/': `${this.APP_BASE}app/`,
       // For test config
@@ -559,7 +560,7 @@ export class SeedConfig {
       '@angular/common/http':
         'node_modules/@angular/common/bundles/common-http.umd.js',
       'tslib': 'node_modules/tslib/tslib.js',
-      'rxjs/operators': 'node_modules/rxjs/operators',
+      'rxjs': 'node_modules/rxjs',
       'dist/tmp/node_modules/*': 'dist/tmp/node_modules/*',
       'node_modules/*': 'node_modules/*',
       '*': 'node_modules/*'
@@ -605,8 +606,24 @@ export class SeedConfig {
         main: 'bundles/service-worker.umd.js',
         defaultExtension: 'js'
       },
-      rxjs: {
-        main: 'Rx.js',
+      'rxjs': {
+        main: 'index.js',
+        defaultExtension: 'js'
+      },
+      'rxjs/ajax': {
+        main: 'index.js',
+        defaultExtension: 'js'
+      },
+      'rxjs/operators': {
+        main: 'index.js',
+        defaultExtension: 'js'
+      },
+      'rxjs/testing': {
+        main: 'index.js',
+        defaultExtension: 'js'
+      },
+      'rxjs/webSocket': {
+        main: 'index.js',
         defaultExtension: 'js'
       }
     }
@@ -712,7 +729,8 @@ export class SeedConfig {
        * @type {object}
        */
       'gulp-sass': {
-        includePaths: ['./node_modules/']
+        includePaths: ['./node_modules/'],
+        importer: magicImporter()
       },
 
       /**
