@@ -15,7 +15,7 @@ export = () => {
   // @angular/service-worker's NGSW CLI
   const sw = join('node_modules', '.bin', 'ngsw-config');
   const src = join('.', 'src', 'client', 'ngsw-config.json');
-  exec(`${sw} ${Config.APP_DEST} ${src}`, function(error: Error, stdout: NodeBuffer, stderr: NodeBuffer) {
+  exec(`${sw} ${Config.APP_DEST} ${src}`, function(error: Error, stdout: any, stderr: any) {
     if (error !== null) {
       reportError('Angular Service Worker config error: ' + error + stderr);
     } else {
